pyproject.toml
[tool.poetry]
name = "russtress-telegram-bot"
version = "0.1.0"
description = ""
authors = ["spdz <spidepog@gmail.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
